requirements.txt
pytorch==2.0.1
torchdiffeq==0.2.3
scanpy==1.9.0
pandas==1.5.0
plotly==5.13.0
jupyterlab==3.6.0
pyyaml==6.0.1
